поправил чтение точек

+++ russ_travel/lib/map/domain/museum_point.dart
@@ -31,8 +31,8 @@ class MuseumPoint extends Equatable {
   factory MuseumPoint.fromJson(Map<String, dynamic> json) {
     return MuseumPoint(
       name: json['name'],
-      latitude: json['latitude'],
-      longitude: json['longitude'],
+      latitude: json['latitude'].toDouble(),
+      longitude: json['longitude'].toDouble(),
       photoUrl: json['photoUrl'], // Получаем ссылку на фото из JSON
     );
   }
